move src files to app/[locale] dir
