+++ frontend/app/[locale]/(auth)/sign-in/[[...sign-in]]/page.tsx
