+++ frontend/app/[locale]/(auth)/sign-up/[[...sign-up]]/page.tsx
