+++ frontend/app/[locale]/(dashboard)/[UserID]/_components/EmptySearch.tsx
