+++ frontend/app/[locale]/(dashboard)/[UserID]/_components/SearchInput.tsx
