+++ frontend/app/[locale]/(dashboard)/[UserID]/_components/board-card/Footer.tsx
