+++ frontend/app/[locale]/(dashboard)/[UserID]/layout.tsx
