+++ frontend/app/[locale]/(dashboard)/[UserID]/page.tsx
