+++ frontend/app/[locale]/(dashboard)/boards/[boardId]/_components/ColorPicker.tsx
