+++ frontend/app/[locale]/(dashboard)/boards/[boardId]/_components/LayerPreview.tsx
