+++ frontend/app/[locale]/(dashboard)/boards/[boardId]/_components/Path.spec.tsx
