+++ frontend/app/[locale]/(dashboard)/boards/[boardId]/_components/Path.tsx
