+++ frontend/app/[locale]/(dashboard)/boards/[boardId]/_components/SelectionTools.tsx
