+++ frontend/app/[locale]/(dashboard)/boards/[boardId]/_components/StylesButton.spec.tsx
