+++ frontend/app/[locale]/(dashboard)/boards/[boardId]/_components/ToolButton.tsx
