+++ frontend/app/[locale]/(dashboard)/boards/[boardId]/page.tsx
